Removed RESULT ATTRIBUTE, caused correlation to go down drastically
After removing the RESULT feature, the correlation dropped from 0.61 to 0.17. 

RESULT column directly correlates with the GN/LS (yards gained lost), because RESULT is the play which led to the yards which were gained/loss

With RESULT being taken out and the correlation dropping, this means the rest of the attributes have no significance in predicting outcome

This may be because the game's history is not being taken into account. In football, each play is determined based on what situation the team is currently in, so game history must be taken into account. 

Therefore, we must now approach this model with the new idea of keeping track of game history.